requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas=1.0=mkl
ca-certificates=2018.03.07=0
certifi=2018.8.13=py36_0
cffi=1.11.5=py36h9745a5d_0
cloudpickle=0.5.3=py36_0
cuda80=1.0=0
cudatoolkit=8.0=3
cudnn=7.0.5=cuda8.0_0
cycler=0.10.0=py_1
dask-core=0.18.2=py36_0
dbus=1.13.0=h3a4f0e9_0
decorator=4.3.0=py36_0
expat=2.2.5=hfc679d8_1
fontconfig=2.13.0=h65d0f4c_5
freetype=2.9.1=h6debe1e_1
gettext=0.19.8.1=0
glib=2.55.0=h464dc38_2
gst-plugins-base=1.12.5=hde13a9d_0
gstreamer=1.12.5=h61a6719_0
icu=58.2=hfc679d8_0
imageio=2.3.0=py36_0
intel-openmp=2018.0.3=0
jpeg=9c=h470a237_0
kiwisolver=1.0.1=py36h2d50403_2
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h470a237_2
libpng=1.6.35=ha92aebf_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.9=he85c1e1_1
libuuid=2.32.1=h470a237_0
libxcb=1.13=h470a237_2
libxml2=2.9.8=h422b904_3
matplotlib=2.2.3=py36h8e2386c_0
mkl=2018.0.3=1
mkl_fft=1.0.4=py36h4414c95_1
mkl_random=1.0.1=py36h4414c95_1
ncurses=6.1=hf484d3e_0
networkx=2.1=py36_0
ninja=1.8.2=py36h6bb024c_1
numpy=1.15.0=py36h1b885b7_0
numpy-base=1.15.0=py36h3dfced4_0
olefile=0.45.1=py36_0
openssl=1.0.2p=h14c3975_0
pcre=8.41=h470a237_2
pillow=5.2.0=py36heded4f4_0
pip=10.0.1=py36_0
pthread-stubs=0.4=h470a237_1
pycparser=2.18=py36_1
pyparsing=2.2.0=py_1
pyqt=5.6.0=py36h8210e8a_6
python=3.6.6=hc3d631a_0
python-dateutil=2.7.3=py_0
pytorch=0.4.1=py36_cuda8.0.61_cudnn7.1.2_1
pytz=2018.5=py_0
pywavelets=0.5.2=py36h035aef0_2
qt=5.6.2=hf70d934_9
readline=7.0=ha6073c6_4
scikit-image=0.14.0=py36hf484d3e_1
scipy=1.1.0=py36hc49cb51_0
setuptools=40.0.0=py36_0
sip=4.18=py36_1
six=1.11.0=py36_1
sqlite=3.24.0=h84994c4_0
tk=8.6.8=0
toolz=0.9.0=py36_0
torchvision=0.2.0=py36_0
tornado=5.1=py36h470a237_1
tqdm=4.24.0=py_1
wheel=0.31.1=py36_0
xorg-libxau=1.0.8=h470a237_6
xorg-libxdmcp=1.1.2=h470a237_7
xz=5.2.4=h14c3975_4
zlib=1.2.11=ha838bed_2
